Damage logic moved to CollisionHandlers. Removed unused events and event listeners. Fixed some maths like lifeTime / maxLifeTime. Optimized Wreck lifeTime calculation. Code refactoring

+++ game-logic/client/src/main/java/net/bfsr/client/renderer/entity/BulletRender.java
@@ -20,7 +20,7 @@ public class BulletRender extends RigidBodyRender<Bullet> {
     public void update() {
         lastPosition.set(object.getPosition());
         lastColor.w = color.w;
-        color.w = 1.0f - object.getLifeTime() / object.getMaxLifeTime();
+        color.w = 1.0f - object.getLifeTime() / (float) object.getMaxLifeTime();
     }
 
     @Override
